refactor(vtc): move leave orchestration into ceremony/ (P2.1, part 2)
The leave `decide → resolve → effect → audit` spine (`remove_inner` +
`assemble_leave_facts` + the `disposition_wire` / `parse_disposition_opt`
helpers) lived in `routes/members/remove.rs` but is shared by three entry points
(self-remove, admin-remove, and the DIDComm self-leave handler in
`messaging.rs`). Move it into `ceremony::orchestrate` beside the role-change
spine (P2.1 part 1, #459).

`remove_inner` now returns a domain `LeaveOutcome { did, disposition, removed }`
rather than the HTTP `RemoveResponse`, keeping the ceremony layer free of wire
types. The route handlers map it via `RemoveResponse::from(outcome)`; the DIDComm
handler maps it into its `SelfRemoveReceiptBody` (already field-by-field). The
member-removed + status-list-flipped audit stays inside the spine — it's the
same for every leave entry point, so it belongs in the orchestration (unlike the
role-change audit, which diverges per caller and stays at the call site).

Behaviour-preserving: removal (15), members_crud (13), ceremony_execute (9),
join_didcomm (1, the DIDComm leave path), full lib (582) all green; fmt + clippy
clean. Join (`submit_inner`) is the remaining spine.
